# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Abrpardaz API",
        "description": "This is the official API which can be used by any user",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1",
            "description": "Local server"
        },
        {
            "url": "http://abrpardaz.com/api/v1",
            "description": "The API endpoint"
        }
    ],
    "paths": {
...
        "/sshKeys/add": {
            "post": {
                "tags": [
                    "Ssh key"
                ],
                "summary": "Add a ssh key to your profile",
                "description": "",
                "operationId": "App\\Http\\Controllers\\V1\\SSHKeyController::add",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/sshKeys/{id}/edit": {
            "post": {
                "tags": [
                    "Ssh key"
                ],
                "summary": "Edit a ssh key using its id",
                "description": "",
                "operationId": "App\\Http\\Controllers\\V1\\SSHKeyController::edit",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/sshKeys/{id}/remove": {
            "delete": {
                "tags": [
                    "Ssh key"
                ],
                "summary": "Remove a ssh key using it",
                "description": "",
                "operationId": "App\\Http\\Controllers\\V1\\SSHKeyController::remove",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/snapshots/ofMachine": {
            "get": {
                "tags": [
                    "Snapshot"
                ],
                "summary": "List snapshots of a specific machine",
                "description": "",
                "operationId": "App\\Http\\Controllers\\V1\\SnapshotController::ofMachine",
                "parameters": [
                    {
                        "name": "machine_id",
                        "in": "query",
                        "description": "id of the machine your want its snapshots",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "List snapshots of a specific machine"
                    }
                }
            }
        },
        "/snapshots/{id}/rename": {
            "post": {
                "tags": [
                    "Snapshot"
                ],
                "summary": "Rename a snapshot",
                "description": "",
                "operationId": "App\\Http\\Controllers\\V1\\SnapshotController::rename",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of the snapshot",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "new name of the snapshot",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "result"
                    }
                }
            }
        },
        "/snapshots/{id}/remove": {
            "delete": {
                "tags": [
                    "Snapshot"
                ],
                "summary": "Removes the snapshot",
                "description": "",
                "operationId": "App\\Http\\Controllers\\V1\\SnapshotController::remove",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of the snapshot",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "result"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login and copy the token into the below box",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}